+++ src/components/barChart-tf-ft.js
@@ -1,7 +1,6 @@
 import { BarChart, Card } from "@tremor/react";
 import { useState } from "react";
 
-
 const dataFormatter = (number) => Intl.NumberFormat("us").format(number).toString();
 
 export default function BarChartFutureCurrent({ tf, years }) {
@@ -12,17 +11,19 @@ export default function BarChartFutureCurrent({ tf, years }) {
     ?.map((year) => {
       return {
         date: year,
-        tf_list: tf?.[year]?.truckFactor,
+        // tf_list: tf?.[year]?.truckFactor,
         future_tf: tf?.[year]?.future_tf,
-        "Current Bus Factor": tf?.[year]?.truckFactor.length,
-        "Potential Bus Factor ": tf?.[year]?.future_tf.length,
+        // "Current Bus Factor": tf?.[year]?.truckFactor.length,
+        [year]: tf?.[year]?.future_tf.length,
       };
     });
 
   return (
     <div className="border p-4">
-      <h3 className="dark:text-dark-tremor-content-strong text-lg font-medium text-tremor-content-strong">Current & Potential Bus Factor</h3>
-      <p className="dark:text-dark-tremor-content text-tremor-default text-tremor-content">The chart below shows the current and potential (future) bus factor</p>
+      <h3 className="dark:text-dark-tremor-content-strong text-lg font-medium text-tremor-content-strong">Potential Bus Factor</h3>
+      <p className="dark:text-dark-tremor-content text-tremor-default text-tremor-content">
+        Evolution of the Potential bus factor over the past {chartData?.length} years
+      </p>
 
       <div className="flex flex-row items-center">
         <BarChart
@@ -31,8 +32,8 @@ export default function BarChartFutureCurrent({ tf, years }) {
           index="date"
           minValue={0}
           allowDecimals={false}
-          categories={["Current Bus Factor", "Potential Bus Factor "]}
-          colors={["blue", "amber", "amber", "rose", "indigo", "emerald"]}
+          categories={years}
+          colors={["amber"]}
           valueFormatter={dataFormatter}
           yAxisWidth={48}
           onValueChange={(v) => setValue(v)}
@@ -45,22 +46,12 @@ export default function BarChartFutureCurrent({ tf, years }) {
 }
 
 const CodeBlock = ({ source }) => {
-  console.log(source?.tf_list);
-
   return (
     <div className="flex min-w-[7rem] flex-col gap-2">
       <div className="text-xs">
-        <p className="font-semibold text-blue-500">Current TF: </p>
-        {source?.tf_list?.map((item, index) => (
-          <span>
-            {index + 1}. {item}
-          </span>
-        ))}
-      </div>
-      <div className="text-xs">
         <p className="font-semibold text-amber-500">Future TF: </p>
         {source?.future_tf?.map((item, index) => (
-          <span>
+          <span key={"ft" + index}>
             {index + 1}. {item}
           </span>
         ))}