+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.0.10] - 2024-06-28
+### Fixed
+- Chart issue and separated the visualization of BF and Future BF
+
+## [1.0.9] - 2024-06-28
+### Fixed
+- Start local server issues
 
 ## [1.0.8] - 2024-06-28
 ### Fixed